demos/04-loan-decision-miniapp/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.5
python-multipart==0.0.20
azure-ai-documentintelligence==1.0.0b4
azure-core==1.32.0
